move(strategy): strategy 패키지로 이동
