+++ src/main/java/car_racing/move_strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package car_racing;
+package car_racing.move_strategy;
 
 public interface MoveStrategy {
     int getMovePoint();
