+++ src/test/java/car_racing/CarTest.java
@@ -1,5 +1,6 @@
 package car_racing;
 
+import car_racing.move_strategy.RandomMoveStopStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
